1.1 Separate Tests from Code
Second major division, between tests and code
